Extract SearchFilters and SearchLayout from GreyVest
