+++ src/index.js
@@ -6,5 +6,7 @@ export * from './utils/schema'
 import FilterAdder from './FilterAdder'
 import ModalFilterAdder from './ModalFilterAdder'
 import FilterButtonList from './FilterButtonList'
+export { default as SearchFilters } from './SearchFilters'
+export { default as SearchLayout } from './SearchLayout'
 
 export { QueryBuilder, FilterAdder, ModalFilterAdder, FilterButtonList }
