+++ stories/imdb/stories/greyVestThemed.js
@@ -4,12 +4,12 @@ import { observable } from 'mobx'
 import { fromPromise } from 'mobx-utils'
 import Contexture, { updateSchemas } from '../utils/contexture'
 import { mergeOverAll } from 'futil-js'
-import { Awaiter, schemaFieldProps, componentForType } from '../../../src'
-import theme, { SearchLayout, SearchFilters, SearchTree, ToggleFiltersHeader, Box, ExampleTypes } from '../../../src/themes/greyVest'
+import { Awaiter, schemaFieldProps, componentForType, SearchLayout, SearchFilters, } from '../../../src'
+import theme, { SearchTree, ToggleFiltersHeader, Box, ExampleTypes } from '../../../src/themes/greyVest'
 import { Column } from './../../../src/layout/ExpandableTable'
 import { ThemeProvider, withTheme } from '../../../src/utils/theme'
 
-let { TypeMap } = ExampleTypes
+let { TypeMap, TermsStatsTable, PagedResultTable, TagsQuery } = ExampleTypes
 
 let tree = Contexture({
   key: 'root',
